FCBHDBP-454 It has done a refactor. It has created a UserAnnotationTrait Trait to handle the same query for notes, bookmarks and highlights.
