+++ app/Models/User/Study/Bookmark.php
@@ -9,7 +9,6 @@ use App\Models\Bible\BibleVerse;
 use App\Models\Bible\Book;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Builder;
 
 /**
  * App\Models\User\Study
@@ -41,6 +40,8 @@ use Illuminate\Database\Eloquent\Builder;
  */
 class Bookmark extends Model
 {
+    use UserAnnotationTrait;
+
     protected $connection = 'dbp_users';
     protected $table = 'user_bookmarks';
     protected $fillable = [
@@ -173,41 +174,5 @@ class Bookmark extends Model
             ->get()
             ->pluck('verse_text');
         return implode(' ', $verses->toArray());
-    }
-
-    /**
-     * Get bookmarks related the playlist items that belong to playlist and a given book ID
-     *
-     * @param Illuminate\Database\Query\Builder $query
-     * @param int $playlist_id
-     * @param string $book_id
-     *
-     * @return Illuminate\Database\Query\Builder
-     */
-    public function scopeWhereBelongPlaylistAndBook(Builder $query, int $playlist_id, string $book_id) : Builder
-    {
-        $dbp_users = config('database.connections.dbp_users.database');
-        $dbp_prod = config('database.connections.dbp.database');
-
-        return $query
-            ->join($dbp_prod . '.bible_fileset_connections AS bfc', 'bfc.bible_id', 'user_bookmarks.bible_id')
-            ->join($dbp_prod . '.bible_filesets AS bf', 'bfc.hash_id', 'bf.hash_id')
-            ->join($dbp_users . '.playlist_items AS pli', function ($join) use ($book_id) {
-                $join
-                    ->on('bf.id', '=', 'pli.fileset_id')
-                    ->where('pli.book_id', $book_id)
-                    ->whereColumn('user_bookmarks.chapter', '=', 'pli.chapter_start')
-                    ->where(function ($wherequery) {
-                        $wherequery
-                        ->orWhereNull('pli.verse_start')
-                        ->orWhere(function ($verse_start_where) {
-                            $verse_start_where
-                                ->whereColumn('user_bookmarks.verse_start', '<=', 'pli.verse_end')
-                                ->whereColumn('user_bookmarks.verse_start', '>=', 'pli.verse_start');
-                        });
-                    });
-            })
-            ->where('pli.playlist_id', $playlist_id)
-            ->where('user_bookmarks.book_id', $book_id);
     }
 }
