+++ app/Models/User/Study/Note.php
@@ -10,7 +10,6 @@ use App\Models\User\User;
 use Illuminate\Contracts\Encryption\DecryptException;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Crypt;
-use Illuminate\Database\Eloquent\Builder;
 
 /**
  * App\Models\User\Note
@@ -38,6 +37,8 @@ use Illuminate\Database\Eloquent\Builder;
  */
 class Note extends Model
 {
+    use UserAnnotationTrait;
+
     protected $connection = 'dbp_users';
     protected $table = 'user_notes';
     protected $hidden = ['user_id'];
@@ -258,45 +259,5 @@ class Note extends Model
         $ctitle = optional($bible->translations->where('language_id', $GLOBALS['i18n_id'])->first())->name;
         $vtitle = optional($bible->vernacularTranslation)->name;
         return ($vtitle ? $vtitle : $ctitle);
-    }
-
-    /**
-     * Get notes related the playlist items that belong to playlist and a given book ID
-     *
-     * @param Illuminate\Database\Query\Builder $query
-     * @param int $playlist_id
-     * @param string $book_id
-     *
-     * @return Illuminate\Database\Query\Builder
-     */
-    public function scopeWhereBelongPlaylistAndBook(Builder $query, int $playlist_id, string $book_id) : Builder
-    {
-        $dbp_users = config('database.connections.dbp_users.database');
-        $dbp_prod = config('database.connections.dbp.database');
-
-        return $query->join($dbp_prod . '.bible_fileset_connections AS bfc', 'bfc.bible_id', 'user_notes.bible_id')
-            ->join($dbp_prod . '.bible_filesets AS bf', 'bfc.hash_id', 'bf.hash_id')
-            ->join($dbp_users . '.playlist_items AS pli', function ($join) use ($book_id) {
-                $join
-                    ->on('bf.id', '=', 'pli.fileset_id')
-                    ->where('pli.book_id', $book_id)
-                    ->whereColumn('user_notes.chapter', '=', 'pli.chapter_start')
-                    ->where(function ($wherequery) {
-                        $wherequery
-                        ->orWhereNull('pli.verse_start')
-                        ->orWhere(function ($verse_start_where) {
-                            $verse_start_where
-                                ->whereColumn('user_notes.verse_start', '<=', 'pli.verse_end')
-                                ->whereColumn('user_notes.verse_start', '>=', 'pli.verse_start');
-                        })
-                        ->orWhere(function ($verse_end_where) {
-                            $verse_end_where
-                                ->whereColumn('user_notes.verse_end', '<=', 'pli.verse_end')
-                                ->whereColumn('user_notes.verse_end', '>=', 'pli.verse_start');
-                        });
-                    });
-            })
-            ->where('pli.playlist_id', $playlist_id)
-            ->where('user_notes.book_id', $book_id);
     }
 }
