+++ app/Transformers/UserNotesTransformer.php
@@ -47,20 +47,19 @@ class UserNotesTransformer extends TransformerAbstract
     public function transform(Note $note)
     {
         return [
-      'id' => (int) $note->id,
-      'bible_id' => (string) $note->bible_id,
-      'bible_name' => (string) $note->bible_name,
-      'book_id' => (string) $note->book_id,
-      'book_name' => (string) optional($note->book)->name,
-      'chapter' => (int) $note->chapter,
-      'verse_start' => (int) $note->verse_start,
-      'verse_end' => (int) $note->verse_end,
-      'verse_text' => (string) $note->verse_text,
-      // 'notes' => (string) $note->notes,
-      'notes' => mb_convert_encoding((string) $note->notes, "UTF-8", "auto"),
-      'created_at' => (string) $note->created_at,
-      'updated_at' => (string) $note->updated_at,
-      'tags' => $note->tags
-    ];
+            'id' => (int) $note->id,
+            'bible_id' => (string) $note->bible_id,
+            'bible_name' => (string) $note->bible_name,
+            'book_id' => (string) $note->book_id,
+            'book_name' => (string) optional($note->book)->name,
+            'chapter' => (int) $note->chapter,
+            'verse_start' => (int) $note->verse_start,
+            'verse_end' => (int) $note->verse_end,
+            'verse_text' => (string) $note->verse_text,
+            'notes' => (string) $note->notes,
+            'created_at' => (string) $note->created_at,
+            'updated_at' => (string) $note->updated_at,
+            'tags' => $note->tags
+        ];
     }
 }
